Move processing own query string and remove Binder
